исправил кредиты, показывается истроия

+++ PrototypeBankSystem.Domain/Entities/Credit.cs
@@ -6,15 +6,16 @@
         public double AmountOfCredit { get; set; }
         public DateTime CreditStart { get; set; }
         public DateTime CreditStop { get; set; }
-        public float InterestRate { get; set; }
+        public double InterestRate { get; set; }
         public bool RepaidLoan { get; set; }
         public int ClientCardID { get; set; }
         public ClientCard ClientCard { get; set; }
+
         public Credit()
         {
         }
 
-        public Credit(int clCardID, double amountOfCredit, DateTime creditStart, DateTime creditStop, float interestRate, bool repaidLoan = false)
+        public Credit(int clCardID, double  amountOfCredit, DateTime creditStart, DateTime creditStop, double interestRate, bool repaidLoan = false)
         {
             ClientCardID = clCardID;
             AmountOfCredit = amountOfCredit;
